feat(HTTPManager): switch to request classes
