--- app/http/requests/index.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const requests = {};
-
-requests.streamStart = require('./streamStart');
-requests.streamStop = require('./streamStop');
-
-module.exports = requests;
