+++ app/http/requests/streaming/index.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const streamingRequests = [];
+
+streamingRequests.push(require('./streamStart'));
+streamingRequests.push(require('./streamStop'));
+
+module.exports = streamingRequests;
